New system test of multiple server cores.

+++ test/wrapper/system_tests/main.sh
@@ -443,6 +443,20 @@ echo "-- PASSED: $test_id"
 rm -rf data/ *.idx
 
 
+test_id="multicore"
+echo "---------------------do $test_id-----------------------"
+python ./multicore/multicore.py $SRCH2_ENGINE_DIR ./multicore/queriesAndResults.txt >> system_test.log 2>&1
+
+if [ $? -gt 0 ]; then
+    echo "FAILED: $test_id"
+    if [ $force -eq 0]; then
+	exit 255
+    fi
+fi
+echo "-- PASSED: $test_id"
+rm -rf data/ *.idx
+
+
 # clear the output directory. First make sure that we are in correct directory
 if [ "$(pwd)" = "$SYSTEM_TEST_DIR" ]; then
     rm -rf data
